LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.arc.selfintersect.1-expected.txt

2d.path.arc.selfintersect.1

arc() with lineWidth > 2*radius is drawn sensibly


PASS arc() with lineWidth > 2*radius is drawn sensibly 
